refactor: Refactored dataset methods
