+++ doctr/datasets/recognition.py
@@ -45,23 +45,3 @@ class RecognitionDataset(AbstractDataset):
             if not isinstance(label, str):
                 raise KeyError("Image is not in referenced in label file")
             self.data.append((img_path, label))
-
-    def __getitem__(
-        self,
-        index: int
-    ) -> Tuple[tf.Tensor, str]:
-
-        img_name, label = self.data[index]
-        img = tf.io.read_file(os.path.join(self.root, img_name))
-        img = tf.image.decode_jpeg(img, channels=3)
-        img = self.sample_transforms(img)
-
-        return img, label
-
-    @staticmethod
-    def collate_fn(samples: List[Tuple[tf.Tensor, str]]) -> Tuple[tf.Tensor, List[str]]:
-
-        images, labels = zip(*samples)
-        images = tf.stack(images, axis=0)
-
-        return images, list(labels)
